requirements.txt
python-docx>=1.2.0
tiktoken>=0.11.0
openai>=1.99.0
python-dotenv>=1.1.0